requirements.txt
albumentations==1.3.0
amqp @ file:///tmp/build/80754af9/amqp_1617383340978/work
aniso8601 @ file:///tmp/build/80754af9/aniso8601_1614708839015/work
attrs @ file:///tmp/build/80754af9/attrs_1620827162558/work
billiard @ file:///C:/ci/billiard_1617383875459/work
cached-property @ file:///tmp/build/80754af9/cached-property_1600785575025/work
celery==5.1.1
certifi @ file:///C:/b/abs_ac29jvt43w/croot/certifi_1665076682579/work/certifi
cffi @ file:///C:/ci/cffi_1613247279197/work
charset-normalizer==2.1.1
click @ file:///home/linux1/recipes/ci/click_1610990599742/work
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
cycler==0.10.0
docopt==0.6.2
dominate==2.6.0
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.10
Flask-HTTPAuth==4.4.0
flask-restplus==0.13.0
flower==1.0.0
fonttools==4.38.0
gevent==21.1.2
greenlet==1.1.0
gunicorn==20.1.0
humanize==3.10.0
idna==3.4
imageio==2.22.4
importlib-metadata==4.6.0
itsdangerous==2.0.1
Jinja2==3.0.1
joblib @ file:///C:/b/abs_e60_bwl1v6/croot/joblib_1666298845728/work
jsonschema==3.2.0
kiwisolver==1.3.1
kombu==5.1.0
MarkupSafe==2.0.1
matplotlib==3.5.3
networkx==2.8.8
nose==1.3.7
numpy==1.21.5
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
packaging==22.0
pafy==0.5.5
pandas==1.3.5
Pillow==9.3.0
pipreqs==0.4.11
prometheus-client==0.11.0
prompt-toolkit==3.0.19
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2021.1
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
redis==3.5.3
requests==2.28.1
Scaffold==0.1.3
scikit-image==0.19.3
scikit-learn==0.23.2
scipy==1.7.3
seaborn==0.12.1
six @ file:///tmp/build/80754af9/six_1644875935023/work
sklearn==0.0
thop==0.1.1.post2209072238
threadpoolctl==2.1.0
tifffile==2022.10.10
torch==1.13.0
torchvision==0.14.0
tornado==6.1
tqdm==4.64.1
typing-extensions==3.10.0.0
urllib3==1.26.13
vine @ file:///Users/ktietz/demo/mc3/conda-bld/vine_1630579390042/work
visitor==0.1.3
wcwidth==0.2.5
Werkzeug==0.16.1
wincertstore==0.2
yarg==0.1.9
zipp==3.4.1
zope.event==4.5.0
zope.interface==5.4.0
